chore: minor improvements, mainly linting fixes (#128)
* chore(readme): fix command in readme

The main `package.json` is under the `frontend` directory and not in repository root. Fix the related `cd` command in the readme.

* chore: improve IDE linting config for tests

Add tsconfigs to the `test` directories to prevent linting errors in VS Code related to global mocha methods like `describe` being undefined. The test sources need to be included in some tsconfig for the TS integration in the IDE to use e.g. the mocha type declarations in the context of the test files.

* chore: fix syntax error in tsconfig.json

* chore: enable linting and type checking for test sources

* test: improve cookie handling utils

The old version of clearCookie() that just sets an empty value to a cookie doesn't actually remove the cookie. It just makes it so the cookie has an empty value but it still exists. To actually remove a cookie it needs to be set with an Expires option in the past.

* test: remove unused const

* test: remove non-functional cookie restoration code

This is not a working way to save and restore cookies using `document.cookie`. This is also not needed since the tests that are setting cookies are also taking care of clearing them using `try..finally` and `clearCookie()`.

* chore: turn off lint rule lit/quoted-expressions

This rule needs to be off for element expressions. Would be nice if there was a rule to enable this only for the other type of expressions like attribute, property and event handler bindings.

* chore: turn off lint rule @typescript-eslint/no-explicit-any

We are currently using explicit any in most source files so let's ignore this for now so we can better focus on not introducing new lint warnings/errors.

* chore: turn off lint rule @typescript-eslint/explicit-module-boundary-types

This gets rid of all the remaining 96 lint warnings.

* test: fix linting issues in tests