feat: add support for re-generating endpoints without restart (#1146)
This feature is based on having either HotSwap Agent
or JRebel set in the development environment for
reloading classes so that Parser and Generator can
re-generate the needed ts files based on the latest
compiled classes.

In this approach, a filesystem WatchService will start
monitoring the classes directory (in a separate thread)
for changes if:

- Application is running in Dev mode, and
- Dev mode live reload is enabled, and
- The backend support for live reload is not Spring Boot DevTools

When a change in the classes is detected, the Parser and
then Generator will update the openapi.json, and since
this file reside under classes itself, then its changes
also detected which will trigger the re-registration of
the Endpoints, without a need to restart the application.

Limitations:
The limitations are dictated by the limitations of JRebel
or HotSwapAgent to redefining the classes at runtime. Some
preliminary local tests shows that HotSwapAgent is working
much smoother than JRebel. For example, JRebel sometimes
fails to redefine the method renames in the memory, although,
Hilla is reloading the change based on the latest changes
completely fine. HowSwapAgent shows a faster reload as well
compared to JRebel.

If Only Spring Boot DevTools is in use in an application,
the watch service will not register for changes, so the
developer experience will not change.

Developer, do not need to do any setup or changes to their
project in order to enable this feature, other than setting
up the HotSwapAgent or JRebel in their development environment.